==> Part_3_RFM with scores.ipynb <==
import pandas as pd
import datetime
import numpy as np

data1 = pd.read_csv('part_2_XYZ_RFM_SALES_final_data_nov_27_invent_class.csv')
data2 = pd.read_csv('part_2_XYZ_RFM_SALES_final_data_nov_27_invent_class.csv')

data2['invoicedate'] = pd.to_datetime(data2['invoicedate'])
data1['invoicedate'] = pd.to_datetime(data1['invoicedate'])

data2.head()

data1.head()

data1['invoicedate'][1]


now = datetime.datetime.now()
print(now)

rfm = data1.groupby('custid', as_index = False).agg({'invoicedate': lambda x: (now - x.max()).days, 'invoiceno': lambda x: len(x.unique()), 'absmargin': lambda x: x.sum(), 'sellprice': lambda x:x.sum()})
rfm['invoicedate'] = rfm['invoicedate'].astype(int)
rfm.rename(columns={'invoicedate': 'recency', 
                         'invoiceno': 'frequency', 
                         'sellprice': 'monetary',
                         'absmargin': 'monetary_margin'}, inplace=True)

rfm.head()

rfm['recency_log'] = np.log(rfm['recency'])
rfm['frequency_log'] = np.log(rfm['frequency'])
rfm['monetary_log'] = np.log(rfm['monetary'])

rfm['monetary_margin_log'] = np.log(rfm['monetary_margin'] + 104578.53418627633 + 1) 

drfm1 = rfm[['custid','recency','frequency_log','monetary_log']]

quantiles = drfm1.quantile(q=[0.25,0.50,0.75])
quantiles

quantiles = quantiles.to_dict()
quantiles

drfm1.head()

segmented_rfm = drfm1

def RScore(x,p,d):
    if x <= d[p][0.25]:
        return 1
    elif x <= d[p][0.50]:
        return 2
    elif x <= d[p][0.75]: 
        return 3
    else:
        return 4
    
def FScore(x,p,d):
    if x <= d[p][0.25]:
        return 4
    elif x <= d[p][0.50]:
        return 3
    elif x <= d[p][0.75]: 
        return 2
    else:
        return 1

segmented_rfm['R'] = segmented_rfm['recency'].apply(lambda x: RScore(x,'recency',quantiles))
segmented_rfm['F'] = segmented_rfm['frequency_log'].apply(lambda x: FScore(x,'frequency_log',quantiles))

segmented_rfm['M'] = segmented_rfm['monetary_log'].apply(lambda x: FScore(x,'monetary_log',quantiles))

segmented_rfm['score'] = segmented_rfm['R'].map(str) + segmented_rfm['F'].map(str) + segmented_rfm['M'].map(str)
segmented_rfm.head()

segmented_rfm.head()

### Customer Category segmentation based on the RFM VALUES

segmented_rfm.loc[(segmented_rfm['R'].isin(['1','2','3','4']))& (segmented_rfm['F'].isin(['1','2'])) & (segmented_rfm['M'].isin(['1','2','3','4'])) ,'category'] = 'Highvalue'
segmented_rfm.loc[(segmented_rfm['R'].isin(['1']))& (segmented_rfm['F'].isin(['1'])) & (segmented_rfm['M'].isin(['1'])) ,'category'] = 'Core'
segmented_rfm.loc[(segmented_rfm['R'].isin(['1','2','3','4']))& (segmented_rfm['F'].isin(['3','4'])) & (segmented_rfm['M'].isin(['1','2','3','4'])) ,'category'] = 'Hibernating'
segmented_rfm.loc[(segmented_rfm['R'].isin(['3','4']))& (segmented_rfm['F'].isin(['1','2','3','4'])) & (segmented_rfm['M'].isin(['1','2'])) ,'category'] = 'At Risk'


try2 = segmented_rfm[['custid','R','F','M','score','category']]

all_added = pd.merge(data2,try2,on='custid',how='left')

all_added.head()

all_added.to_csv('part_3_RFM_SCORE_sales_data_with_customer_RFM_score_nov_27_final.csv',index=False)
