==> Part_4_Feature development for SCORING MODEL and Determining Customer performance GOOD or BAD.ipynb <==
import pandas as pd

final_data = pd.read_csv('part_3_RFM_SCORE_sales_data_with_customer_RFM_score_nov_27_final.csv')

final_data.head()


## Feature development 

#### For a customer id: Final  invent_class is the maxmimum frequency a particular invent class was order

#### Feature 1: Invent_class per customer_id

t1 = final_data.groupby(['custid','invent_class'],as_index=False).agg({'quantity': lambda x: x.sum()})

t1.head()

t2 = t1.groupby('custid',as_index=False).agg({'quantity': lambda x: x.max()})
t2.rename(columns={'quantity':'max_quantity'},inplace=True)

t2.head()

class_dist = pd.merge(t2,t1,how='left',left_on=['custid','max_quantity'],right_on=['custid','quantity'])

final_invent_class_label = class_dist[['custid','invent_class']]

len(final_invent_class_label)

## Sorting  dataframe

final_invent_class_label.sort_values(by='custid',inplace=True)

final_invent_class_label.invent_class.value_counts()

#### Feature 2: Product portfolio: Out of total 5 commodities, which customer orders maximum type of  commodities

product_porfolio = final_data.groupby(['custid'],as_index=False).agg({'commodity': lambda x:  len(x.unique())})
product_porfolio.rename(columns={'commodity':'product_portfolio'},inplace=True)

len(product_porfolio)

## Sorting dataframe
product_porfolio.sort_values(by='custid',inplace=True)

#### Feature 3: History_length , Customer association with the company

final_data['invoicedate'] = pd.to_datetime(final_data['invoicedate'])

final_data['mix_date']=final_data['invoicedate']
final_data['max_date']=final_data['invoicedate']

dates_data=final_data.groupby('custid',as_index=False).agg({'mix_date': lambda x: x.min(),'max_date': lambda x: x.max()})

dates_data['length_history'] = (dates_data['max_date']-dates_data['mix_date']).dt.days

len(dates_data)

dates_data[dates_data['length_history']>30].count()

quantiles_history = dates_data['length_history'].quantile(q=[0.25,0.50,0.75])

def hist_score(x):
    if x <= quantiles_history[0.25]:
        return 4
    elif x <= quantiles_history[0.50]:
        return 3
    elif x <= quantiles_history[0.75]:
        return 2
    else:
        return 1


dates_data['length_history'] = dates_data['length_history'].apply( lambda x: hist_score(x))

len(dates_data)

## Sorting dataframe
dates_data.sort_values(by='custid',inplace=True)

##### Feature 4: R F M

rfm_date = final_data.groupby(['custid'],as_index=False).agg({'R': 'mean','F': 'mean','M': 'mean'})

len(rfm_date)

## Sorting dataframe
rfm_date.sort_values(by='custid',inplace=True)

## Merging all feature together 

final1_2 = pd.merge(dates_data,rfm_date,on='custid',how='left')

final1_2.head()

final1_2 = pd.merge(final1_2,product_porfolio,on='custid',how='left')

final1_2.head()

final1_2 = pd.merge(final1_2,final_invent_class_label,on='custid',how='left')

final = final1_2[['custid','length_history','R','F','M','product_portfolio','invent_class']]

final.head()

final.to_csv('part_4_Customer_pre_final_NOV27.csv',index=False)

###  weightage for attributes in individual features
#### IF R = 1 , weightage for attribut R1 = 1
#### similary if R=2 , weightage for attribute R2 = 0.8
#### and so on 

#### 'R': {1: 1.0, 2: 0.8, 3: 0.5, 4: 0.1}
    
#### 'F': {1: 1.0, 2: 0.8, 3: 0.5,4: 0.1}
 
#### 'M': {1: 1.0, 2: 0.8, 3: 0.5,4: 0.1}

#### 'product_portfolio': {5: 1.0, 4: 0.9, 3: 0.8, 2: 0.7, 1 : 0.6}
    
#### 'invent_class': {'green':1.0, 'yellow': 0.8 ,'red': 0.5}
    
#### 'length_history':{1: 1.0, 2: 0.8, 3: 0.6, 4: 0.4}


final_labels = {'R': {1: 1.0, 2: 0.8, 3: 0.5, 4: 0.1},
 'F': {1: 1.0, 2: 0.8, 3: 0.5,4: 0.1},
 'M': {1: 1.0, 2: 0.8, 3: 0.5,4: 0.1},
 'product_portfolio': {5: 1.0, 4: 0.9, 3: 0.8, 2: 0.7, 1 : 0.6},
 'invent_class': {'green':1.0, 'yellow': 0.8 ,'red': 0.5},
 'length_history':{1: 1.0, 2: 0.8, 3: 0.6, 4: 0.4}}




def final_scaling(x,key1):
    weights={'R':0.15,'M':0.25,'F':0.25,'product_portfolio':0.15,'invent_class':0.10,'length_history':0.10}
    current_weight = weights[key1]
    return current_weight*(final_labels[key1][x])

final_weighted = final.copy()

final_weighted.product_portfolio.isnull().sum()

final_weighted['F_weighted'] = final_weighted['F'].apply(final_scaling,args=('F'))
final_weighted['M_weighted'] = (final_weighted['M'].apply(final_scaling,args=('M')))
final_weighted['R_weighted'] = (final_weighted['R'].apply(final_scaling,args=('R')))
# product_portfolio = (trial3['product_portfolio'].apply(final_scaling,args=['product_portfolio']))
# invent_class = (trial3['invent_class'].apply(final_scaling,args=('invent_class')))
# length_history = (trial3['length_history'].apply(final_scaling,args=('length_history')))

# final['R_factored'] = final['R'].apply(lambda x: final_scaling('R',x))
# final['M_factored'] = final['M'].apply(lambda x: final_scaling('M',x))
# final['factored_product_portfolio'] = final['product_portfolio'].apply(lambda x: final_scaling('product_portfolio',x))
# final['factored_invent_class'] = final['invent_class'].apply(lambda x: final_scaling('invent_class',x))
# final['length_history'] = final['length_history'].apply(lambda x: final_scaling('length_history',x))

final_weighted['product_portfolio_weighted']= final_weighted['product_portfolio'].apply(lambda x: final_scaling(x,'product_portfolio'))
final_weighted['invent_class_weighted']= final_weighted['invent_class'].apply(lambda x: final_scaling(x,'invent_class'))
final_weighted['length_history_weighted'] = final_weighted['length_history'].apply(lambda x: final_scaling(x,'length_history'))


final_weighted.head()

final_weighted['final_value'] = final_weighted['product_portfolio_weighted']+final_weighted['invent_class_weighted']+final_weighted['length_history_weighted']+final_weighted['F_weighted']+final_weighted['M_weighted']+final_weighted['R_weighted']

final_weighted.head()

#### Mapping Final value 
#### IF final value os <=0.5 = 0 or Bad performance
#### and IF value >0.5  map it to 1 or Good performance

def performance_map(x):
    if x<=0.6:
        response = 0 #'Bad'
        return response
    else:
        response = 1# 'Good'
        return response

final_weighted['customer_type'] = final_weighted['final_value'].map(performance_map)

#### Distribution of GOOD and Bad

final_weighted.customer_type.value_counts()

len('product_portfolio')

final_weighted.head()

## Exporting File with 

final_weighted.to_csv('part_4_Customer_GOOD_BAD_parameter_nov_final.csv',index=False)
