==> Part_5_Final Scoring Model.ipynb <==
import pandas as pd

final_data = pd.read_csv('part_4_Customer_GOOD_BAD_parameter_nov_final.csv')

### Preparing data for Logistic regression

final_data.head()

test1 = final_data[['length_history','R','F','M','product_portfolio','invent_class']]

test1.head()

#### Mapping all parametes to string type

test1['length_history'] = test1['length_history'].map(str)
test1['R'] = test1['R'].map(str)
test1['F'] = test1['F'].map(str)
test1['M'] = test1['M'].map(str)
test1['product_portfolio'] = test1['product_portfolio'].map(str)


### Getting Dummies for dataset
X=pd.get_dummies(test1)
y=final_data['customer_type']



X.head()

from sklearn.linear_model import LogisticRegression
from sklearn import metrics
logreg = LogisticRegression()
logreg.fit(X, y)

((logreg.coef_)[0])

logreg.intercept_

columns = X.columns

mn=list(zip(X.columns,((logreg.coef_)[0])))

test = pd.DataFrame(mn)

test.rename(columns={0:'Columns',1:'Coefficients'},inplace=True)

test1 = test.append({'Columns' : 'Intercept' , 'Coefficients' :   0.69115405} , ignore_index=True)

### 1.72 is Intercept which comes from logreg.intercept_

def root_extract(x):
    if len(x)>6:
        return x[:5]
    else:
        return x[:1]

test1['Root'] = test1['Columns'].apply(lambda x: root_extract(x))

test1

min_coef_sum = test1.groupby('Root').min().sum()

min_coef_sum =(min_coef_sum).reset_index().iloc[1,1]

max_coef_sum = test1.groupby('Root').max().sum()

max_coef_sum=(max_coef_sum).reset_index().iloc[1,1]

#### OUR custom scoring range

max_score = 850
min_score = 300


def scale_value(x):
    new_value = x*((max_score-min_score)/(max_coef_sum-min_coef_sum))
    return new_value

test1['Coefficients']=test1['Coefficients'].astype('float')

test1['scaled_value']=test1['Coefficients'].apply(lambda x: scale_value(x))

test1.info()

test1['scaled_value_copy'] = test1['scaled_value']

intercept_value_scaled = ((( 0.69115405- min_coef_sum)/(max_coef_sum-min_coef_sum))*(max_score-min_score))+min_score

### Intercept scaled value 
intercept_value_scaled

test1.iloc[24,4] =intercept_value_scaled

#### SCORING BASED ON INDIVIDUAL attributes :: Scaled_value_copy final scores per attribute

test1

test1.groupby('Root').max().sum()

test1.rename(columns={'scaled_value_copy':'scaled_value_intercept_adjust'},inplace=True)


scaled_score_customer = test1.iloc[:,4]

scaled_score_customer

X.head()

data_with_vaiable=X.astype(int)

data_with_vaiable['customer_performace']= y

data_with_vaiable.head()

data_with_vaiable['custid']= final_data['custid']

data_with_vaiable.head()

#### Re-arranging column in the dataframe

columns = data_with_vaiable.columns.tolist()

columns = columns[-1:] + columns[:-1]

Data_with_dummy_variable =data_with_vaiable[columns]

Data_with_dummy_variable['custid'] = Data_with_dummy_variable['custid'].astype(str)

Data_with_dummy_variable.head()

final_Score_tes1 = Data_with_dummy_variable.copy()

#### Function for generating Customer score based on individual attribute scores

def final_score(x):
    """ intrercept: scaled_score_customer[24] 
    other score """
    score = scaled_score_customer[24] + scaled_score_customer[0]*x[1]+ scaled_score_customer[1]*x[2]+ scaled_score_customer[2]*x[3]+ scaled_score_customer[3]*x[4]+ scaled_score_customer[4]*x[5]\
    + scaled_score_customer[5]*x[6]+ scaled_score_customer[6]*x[7]+ scaled_score_customer[7]*x[8]+ scaled_score_customer[8]*x[9]+ scaled_score_customer[9]*x[10]+ scaled_score_customer[10]*x[11]\
    + scaled_score_customer[11]*x[12]+ scaled_score_customer[12]*x[13]+ scaled_score_customer[13]*x[14]+ scaled_score_customer[14]*x[15]+ scaled_score_customer[15]*x[16]\
    + scaled_score_customer[16]*x[17]+ scaled_score_customer[17]*x[18]+ scaled_score_customer[18]*x[19]+ scaled_score_customer[19]*x[20]+ scaled_score_customer[20]*x[21]\
    + scaled_score_customer[21]*x[22]+ scaled_score_customer[22]*x[23]+ scaled_score_customer[23]*x[24]
    
    return score

#### Passing complete row in the apply function 

final_Score_tes1['cust_score']=final_Score_tes1.apply(lambda row : final_score(row), axis = 1) 

final_Score_tes1['cust_score']= (round(final_Score_tes1['cust_score'])).astype(int)

final_Score_tes1.head()

## Intercept value
scaled_score_customer[24]

## Minimum score 
final_Score_tes1['cust_score'].min()

## Maximum score 
final_Score_tes1['cust_score'].max()

data_with_scorecard = final_Score_tes1[['custid','cust_score']]

new_subset_1 = final_data[['custid','R','F','M','product_portfolio','invent_class','length_history','customer_type']]


new_subset_1['custid'] =new_subset_1['custid'].astype(str)

new_subset_1['RFM_Score'] = new_subset_1['R'].map(str) + new_subset_1['F'].map(str) + new_subset_1['M'].map(str)

### Customer category based on RFM VALUES, 


new_subset_1.loc[(new_subset_1['R'].isin(['1','2','3','4']))& (new_subset_1['F'].isin(['1','2'])) & (new_subset_1['M'].isin(['1','2','3','4'])) ,'category'] = 'Highvalue'
new_subset_1.loc[(new_subset_1['R'].isin(['1']))& (new_subset_1['F'].isin(['1'])) & (new_subset_1['M'].isin(['1'])) ,'category'] = 'Core'
new_subset_1.loc[(new_subset_1['R'].isin(['1','2','3','4']))& (new_subset_1['F'].isin(['3','4'])) & (new_subset_1['M'].isin(['1','2','3','4'])) ,'category'] = 'Hibernating'
new_subset_1.loc[(new_subset_1['R'].isin(['3','4']))& (new_subset_1['F'].isin(['1','2','3','4'])) & (new_subset_1['M'].isin(['1','2'])) ,'category'] = 'At Risk'


final_data_with_scorecard = pd.merge(new_subset_1,data_with_scorecard,on='custid',how='left')

final_data_with_scorecard.head()

### Mapping Customer Rating 

def map_score(x):
    if x<=450:
        return 'Very Poor'
    elif  x>450 and x<=620:
        return 'Fair'
    elif x>620 and x<=700:
        return 'Good'
    elif x>700 and x<=799:
        return 'Very Good'
    else:
        return 'Exceptional'

final_data_with_scorecard['rating']= final_data_with_scorecard['cust_score'].map(map_score)

final_data_with_scorecard.head()

### Distribution of customer rating

# final_data_with_scorecard['rating'].value_counts()/final_data_with_scorecard['rating'].count() 
final_data_with_scorecard['rating'].value_counts(normalize=True)

final_data_with_scorecard.to_csv('part_5_Customer_Final_DATA_WITH_SCORES.csv',index=False)

import pandas as pd

df = pd.read_csv('part_5_Customer_Final_DATA_WITH_SCORES.csv')

df.cust_score.mean()
